Moved RMSIcon to Components folder
